requirements.txt
pytest~=8.3.2
selenium~=4.23.1
Faker~=26.0.0
requests~=2.32.3
allure-pytest